refactor: add custom hook for form
